Safari-609.1.11.1/LayoutTests/fast/canvas/2d.getPath.newobject-expected.txt

The test to ensure that each call to getPath returns a new object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS path1 is not path2
PASS path1.currentX is path2.currentX
PASS path1.currentY is path2.currentY
PASS successfullyParsed is true

TEST COMPLETE
